requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
pytimeparse==1.1.8